make data privacy checkbox reusable component
